+++ src/uptime/uptime.rs
@@ -55,7 +55,7 @@ unsafe extern fn utmpxname(_file: *const c_char) -> c_int {
     0
 }
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("v", "version", "output version information and exit"),
